Move punch::Enemy to a separate module
