+++ src/engine/punch/mod.rs
@@ -1,6 +1,8 @@
 pub use self::scene::Scene;
 pub use self::player::Player;
+pub use self::enemy::Enemy;
 
 pub mod scripting;
+mod enemy;
 mod player;
 mod scene;
